validated the project and increased the hight

+++ src/main/java/lk/ijse/oxford/contoller/classroom_form_controllers/DeleteClassroomFormController.java
@@ -1,17 +1,21 @@
 package lk.ijse.oxford.contoller.classroom_form_controllers;
 
+import com.jfoenix.controls.JFXTextField;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
 import lk.ijse.oxford.model.Classroom;
 import lk.ijse.oxford.model.tm.ClassTm;
 import lk.ijse.oxford.model.tm.EquipmentTm;
 import lk.ijse.oxford.repository.ClassroomRepo;
 import lk.ijse.oxford.repository.EquipmentRepo;
+import lk.ijse.oxford.util.Regex;
+import lk.ijse.oxford.util.TextFields;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -25,9 +29,9 @@ public class DeleteClassroomFormController {
     private Label lblClassCapacity;
     private int classCapacity;
     @FXML
-    private TextField txtClassId;
+    private JFXTextField txtClassId;
     @FXML
-    private TextField txtClassDesc;
+    private JFXTextField txtClassDesc;
     @FXML
     private TableColumn<?,?> colClassId;
     @FXML
@@ -95,17 +99,20 @@ public class DeleteClassroomFormController {
         colSubId.setCellValueFactory(new PropertyValueFactory<>("subId"));
     }
     public void btnClassDeleteOnAction(ActionEvent actionEvent) {
-        String id = txtClassId.getText();
+        if (isValidate()){
+            String id = txtClassId.getText();
 
-        try {
-            boolean isDeleted = ClassroomRepo.delete(id);
-            if (isDeleted) {
-                new Alert(Alert.AlertType.CONFIRMATION, "Classroom Data Deleted!").show();
-                txtClassDesc.setText("");
-                txtClassId.setText("");
+            try {
+                boolean isDeleted = ClassroomRepo.delete(id);
+                if (isDeleted) {
+                    new Alert(Alert.AlertType.CONFIRMATION, "Classroom Data is Deleted!").show();
+                    initialize();
+                    txtClassDesc.setText("");
+                    txtClassId.setText("");
+                }
+            } catch (SQLException e) {
+                new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
             }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
         }
     }
 
@@ -115,9 +122,16 @@ public class DeleteClassroomFormController {
         txtClassDesc.setText(selectedItem.getDescription());
     }
 
-    public void btnClassRefreshOnAction(ActionEvent actionEvent) {
-        initialize();
-        txtClassDesc.setText("");
-        txtClassId.setText("");
+    public void txtDescCheckOnAction(KeyEvent keyEvent) {
+        Regex.setTextColor(TextFields.DESC,txtClassDesc);
+    }
+
+    public void txtIdCheckOnAction(KeyEvent keyEvent) {
+        Regex.setTextColor(TextFields.CID,txtClassId);
+    }
+    public boolean isValidate(){
+        if(!Regex.setTextColor(TextFields.CID,txtClassId))return false;
+        if(!Regex.setTextColor(TextFields.DESC,txtClassDesc))return false;
+        return true;
     }
 }
