+++ src/main/java/lk/ijse/oxford/contoller/equipment_form_controllers/DeleteEquipmentFormController.java
@@ -1,16 +1,20 @@
 package lk.ijse.oxford.contoller.equipment_form_controllers;
 
+import com.jfoenix.controls.JFXTextField;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
 import lk.ijse.oxford.model.Equipment;
 import lk.ijse.oxford.model.tm.EmployeeTm;
 import lk.ijse.oxford.model.tm.EquipmentTm;
 import lk.ijse.oxford.repository.EquipmentRepo;
+import lk.ijse.oxford.util.Regex;
+import lk.ijse.oxford.util.TextFields;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -21,9 +25,9 @@ public class DeleteEquipmentFormController {
     private Label lblEquipCount;
     private int equipCount;
     @FXML
-    private TextField txtEquipId;
+    private JFXTextField txtEquipId;
     @FXML
-    private TextField txtEquiptDesc;
+    private JFXTextField txtEquiptDesc;
     @FXML
     private TableColumn<?,?> colEquipmentId;
     @FXML
@@ -83,16 +87,19 @@ public class DeleteEquipmentFormController {
         colEquipmentQty.setCellValueFactory(new PropertyValueFactory<>("qty"));
     }
     public void btnEquipmentDeleteOnAction(ActionEvent actionEvent) {
-        String id = txtEquipId.getText();
+       if (isValidate()){
+           String id = txtEquipId.getText();
 
-        try {
-            boolean isDeleted = EquipmentRepo.delete(id);
-            if (isDeleted) {
-                new Alert(Alert.AlertType.CONFIRMATION, "Equipment Data Deleted!").show();
-            }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
-        }
+           try {
+               boolean isDeleted = EquipmentRepo.delete(id);
+               if (isDeleted) {
+                   new Alert(Alert.AlertType.CONFIRMATION, "Equipment Data Deleted!").show();
+                   initialize();
+               }
+           } catch (SQLException e) {
+               new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
+           }
+       }
     }
 
     public void equipDeleteOnAction(MouseEvent mouseEvent) {
@@ -101,7 +108,16 @@ public class DeleteEquipmentFormController {
         txtEquiptDesc.setText(selectedItem.getDescription());
     }
 
-    public void btnEmpRefresh(ActionEvent actionEvent) {
-        initialize();
+    public boolean isValidate(){
+        if(!Regex.setTextColor(TextFields.DESC,txtEquiptDesc))return false;
+        if(!Regex.setTextColor(TextFields.EQID,txtEquipId))return false;
+        return true;
+    }
+    public void txtClassDescCheckOnAction(KeyEvent keyEvent) {
+        Regex.setTextColor(TextFields.DESC,txtEquiptDesc);
+    }
+
+    public void txtClassIdCheckOnAction(KeyEvent keyEvent) {
+        Regex.setTextColor(TextFields.EQID,txtEquipId);
     }
 }
